chore(email-erification):finish email verification [Finish #167730740]
