+++ src/utils/index.js
@@ -9,4 +9,6 @@ const emailVerifyToken = id => new Promise((resolve, reject) => {
 	if (!token) reject({ err: 'could not assign a token, make sure you provide a secret key' });
 });
 
+
+
 export { emailVerifyToken };
